Make list fingerprint(s) intent SDM specific
Prior to this commit all SDMs with the fingerprint pack support
registered used the same intent names for listing fingerprints.  This
caused SDMs to override each other's intents.  This commit adds the SDM
name as a suffix to the intent to keep it unique (as has been done to
the target commands).